+++ Admin/notifier/js/notifier.js
@@ -67,6 +67,13 @@ floater.onclick = function() {
                 translateY: 0,
                 opacity: 1,
               })
+
+              floater.onclick = function() {
+
+                modal.alert('Select user to send', () => {
+                  userboxUI.create('Name', 'subname', 'https://api.dicebear.com/9.x/shapes/svg?seed=Bjrt&radius=50&size=58', userboxUI.input('648'))
+                }, '')
+              }
             }
           })
         }